Prometheus Sink review comment rsolved for issue #1744. Signed-off-by: mallikagogoi7 <mallikagogoi7@gmail.com>
