+++ data-prepper-plugins/prometheus-sink/src/main/java/org/opensearch/dataprepper/plugins/sink/prometheus/configuration/AuthTypeOptions.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.dataprepper.plugins.sink.configuration;
+package org.opensearch.dataprepper.plugins.sink.prometheus.configuration;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 
@@ -10,11 +10,12 @@ import java.util.Arrays;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-public enum HTTPMethodOptions {
-    PUT("PUT"),
-    POST("POST");
+public enum AuthTypeOptions {
+    HTTP_BASIC("http-basic"),
+    BEARER_TOKEN("bearer-token"),
+    UNAUTHENTICATED("unauthenticated");
 
-    private static final Map<String, HTTPMethodOptions> OPTIONS_MAP = Arrays.stream(HTTPMethodOptions.values())
+    private static final Map<String, AuthTypeOptions> OPTIONS_MAP = Arrays.stream(AuthTypeOptions.values())
             .collect(Collectors.toMap(
                     value -> value.option,
                     value -> value
@@ -22,12 +23,12 @@ public enum HTTPMethodOptions {
 
     private final String option;
 
-    HTTPMethodOptions(final String option) {
+    AuthTypeOptions(final String option) {
         this.option = option;
     }
 
     @JsonCreator
-    static HTTPMethodOptions fromOptionValue(final String option) {
+    static AuthTypeOptions fromOptionValue(final String option) {
         return OPTIONS_MAP.get(option);
     }
 }
