+++ data-prepper-plugins/prometheus-sink/src/main/java/org/opensearch/dataprepper/plugins/sink/prometheus/service/PrometheusSinkAwsService.java
@@ -2,11 +2,11 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.dataprepper.plugins.sink.service;
+package org.opensearch.dataprepper.plugins.sink.prometheus.service;
 
 import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
 import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
-import org.opensearch.dataprepper.plugins.sink.configuration.PrometheusSinkConfiguration;
+import org.opensearch.dataprepper.plugins.sink.prometheus.configuration.PrometheusSinkConfiguration;
 
 /**
  * This service class contains logic for handling AWS related services
